repeticao e operadores

+++ src/calculadora/Calculadora.java
@@ -6,25 +6,29 @@ public class Calculadora {
         System.out.println("Resultado da sua soma é: " + resultado);
         return resultado;
     }
+
     public float subtrai(float a, float b) {
         float resultado = a - b;
         System.out.println("Resultado da sua subtracao é: " + resultado);
         return resultado;
     }
+
     public float multiplica(float a, float b) {
         float resultado = a * b;
         System.out.println("Resultado da sua multiplicacao é: " + resultado);
         return resultado;
     }
+
     public float divide(float a, float b) {
         float resultado = a / b;
         System.out.println("Resultado da sua divisao é: " + resultado);
         return resultado;
     }
+
     public float porcentagem(float a, float b) {
-        float c = (float) a / b;
+        float c = a / b;
         float resultado = c * 100;
-        System.out.println(a+" é "+resultado+"% de "+b);
+        System.out.println(a + " é " + resultado + "% de " + b);
         return resultado;
 
     }
